program C imprime la taille de différents types sur l'ordinateur sur lequel il est compilé et exécuté

+++ hello_world/6-size.c
@@ -14,11 +14,11 @@ long int c;
 long long int d;
 float f;
 
-printf("Size of char: %lu byte(s)\n", sizeof(a));
-printf("Size of int: %lu byte(s)\n", sizeof(b));
-printf("Size of long int: %lu byte(s)\n", sizeof(c));
-printf("Size of long long ing: %lu byte(s)\n", sizeof(d));
-printf("Size of float: %lu byte(s)\n", sizeof(f));
+printf("Size of char: %lu byte(s)\n", (unsigned long)sizeof(a));
+printf("Size of int: %lu byte(s)\n", (unsigned long)sizeof(b));
+printf("Size of long int: %lu byte(s)\n", (unsigned long)sizeof(c));
+printf("Size of long long ing: %lu byte(s)\n", (unsigned long)sizeof(d));
+printf("Size of float: %lu byte(s)\n", (unsigned long)sizeof(f));
 
 return (0);
 }
